refactor: rename graphql folder → graphql schema
