+++ packages/takudev-web/graphql_schema/codegen.yaml
@@ -1,12 +1,12 @@
 overwrite: true
-schema: ./graphql/schema.graphql
-documents: ./graphql/documents/**/*.graphql
+schema: ./graphql_schema/schema.graphql
+documents: ./graphql_schema/documents/**/*.graphql
 generates:
-  ./graphql/__generated__/index.d.ts:
+  ./graphql_schema/__generated__/index.d.ts:
     plugins:
       - typescript
       - typescript-operations
-  ./graphql/__generated__/index.ts:
+  ./graphql_schema/__generated__/index.ts:
     preset: import-types
     presetConfig:
       typesPath: ./index.d
