+++ packages/takudev-web/graphql_schema/documents/query/findArticleById.graphql
