+++ packages/takudev-web/graphql_schema/documents/query/getAllArticleSummary.graphql
