+++ packages/takudev-web/pages/articles/[slug].tsx
@@ -4,7 +4,7 @@ import Head from 'next/head';
 import React from 'react';
 import isEqual from 'react-fast-compare';
 import { PageTemplate } from '~/components/templates';
-import { getGraphqlSdk } from '~/graphql';
+import { getGraphqlSdk } from '~/graphql_schema';
 import {
   isNotNullable,
   filterNotNullableElement,
