+++ packages/takudev-web/pages/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import isEqual from 'react-fast-compare';
 import { HomePageContent } from '~/components/features';
 import { PageTemplate } from '~/components/templates';
-import { getGraphqlSdk } from '~/graphql';
+import { getGraphqlSdk } from '~/graphql_schema';
 import {
   isNotNullable,
   filterNotNullableElement,
